Add an output buffer to MachineBase classes
Output buffer of size one was added to MachineBase to ensure that a machine can
receive a part at the same timestamp as when it finishes processing last part.
This means Machine class can hold up to 2 parts (input and putput) but the new
input part will not begin to be processed until the output buffer is empty.