Rework Maintainer class.
Update Maintainer to do work orders which are more generic. Maintainer
now calls the target when work order starts and when it finishes.
Add Maintainable class which needs to be subclassed by classes for
Maintainer to be able to perform work orders on those classes.
Changed assets not to record changes in value when the change is zero.