Add a warning to DecisionGate.
DecisionGate should only use the state of the Part being passed to determine
if it is allowed to pass because using other information may break Part flow.
DecisionGate does not check whether a Part can pass unless a Part is being
passed meaning if the DecisionGate blocks a Part and later would allow the same
Part to pass there is nothing to trigger the attempt to pass that Part.